Detecting CJSON directories
